requirements.txt
# EXFIN_API Requirements
# Python 3.9+

# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database Drivers
psycopg2-binary==2.9.9
pymssql==2.2.8

# Security & Authentication
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Logging
loguru==0.7.2

# Data Validation
pydantic
pydantic-settings

# Dashboard / Reporting (Streamlit)
streamlit
streamlit-authenticator
pandas
plotly
sqlalchemy
requests
pyodbc
openai
pyarrow
fastparquet
# PDF Generation
# PDF Generation
reportlab
Pillow

# WhatsApp Integration (Twilio)
twilio==8.10.0

# HTTP Requests
requests==2.31.0

# Date & Time
python-dateutil==2.8.2
APScheduler==3.10.4
SQLAlchemy==2.0.23


# GUI / System Tray

# Template Engine
jinja2==3.1.2

# System & Process Management
tendo>=0.3.0
pywin32>=306
psutil>=5.9.0
asyncpg
bcrypt
openpyxl
redis
email-validator
